added pathing support for windows, fixed a write file issue in changeLog.js
