+++ prompts/index.js
@@ -1,0 +1,7 @@
+const settupPrompts = require("./settupPrompts");
+const basePrompts = require("./basePrompts");
+
+module.exports = {
+    settupPrompts,
+    basePrompts,
+}
