+++ utils/changeLog.js
@@ -12,10 +12,11 @@ class ChangeLog{
                 if (err) return console.log(err);
                 //console.log('writing to ' + "./utils/changeLog.json");
             });
+            return changeLog.changesToPush;
         }
     }
     updateLog(changes) {
-        const changeLog = this.getLog();
+        const changeLog = {};
         changeLog.changesToPush = changes;
         fs.writeFile("./utils/changeLog.json", JSON.stringify(changeLog,null,2), function writeJSON(err) {
             if (err) return console.log(err);
@@ -23,8 +24,10 @@ class ChangeLog{
         });
     }
     pushToLog(change) {
-        const changeLog = this.getLog();
-        changeLog.changesToPush.push(change);
+        const changeLog = {
+            changesToPush: [...this.getLog(),change]
+        };
+        // changeLog.changesToPush.push(change);
         fs.writeFile("./utils/changeLog.json", JSON.stringify(changeLog,null,2), function writeJSON(err) {
             if (err) return console.log(err);
             //console.log('writing to ' + "./utils/changeLog.json");
