+++ utils/constants.js
@@ -1,0 +1,16 @@
+const EXIT = "\033[91mExit\x1b[0m";
+const BACK = "\033[33mBack\x1b[0m";
+
+const ERROR_COLOR = "\x1b[31m%s\x1b[0m";
+const WARNING_COLOR = "\x1b[33m%s\x1b[0m";
+const SUCCESS_COLOR = "\x1b[32m%s\x1b[0m";
+const INFO_COLOR = "\x1b[34m%s\x1b[0m";
+
+module.exports = {
+    EXIT,
+    BACK,
+    ERROR_COLOR,
+    WARNING_COLOR,
+    SUCCESS_COLOR,
+    INFO_COLOR,
+}
